requirements/main.txt
diff_match_patch
editdistance
ffmpeg
hydra-core
joblib
librosa>=0.10.0 # specify >=0.10.0 so that librosa.get_duration(path=...) will work
numpy
omegaconf
pandas
rarfile
regex
sox
tqdm
webvtt-py
wget
pydub

# for some processers, additionally https://github.com/NVIDIA/NeMo is required
# for some processers, additionally nemo_text_processing is required
# for mcv: apt-get update && apt-get upgrade -y && apt-get install -y sox libsox-fmt-all
